move all watcher related code to the swap_watcher module
